render: от элементов, к компонентам
