+++ src/util.js
@@ -1,9 +1,3 @@
-export const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`
-};
-
-
 export const getRandomArrayItems = (array, maxItems = array.length, minItems = 1, joinSTR = `, `) => {
   const arr = array.slice();
   const itemsAmount = getRandomIntegerNumber(minItems, maxItems);
@@ -41,22 +35,4 @@ export const getAmountByCurrentKey = (objects, key, value) => {
 export const sortObjectsByKeyMaxMin = (objects, key) => {
   const arr = objects.slice();
   return arr.sort((a, b) => b[key] - a[key]);
-};
-
-export const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
-
-export const render = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-  }
 };
